extract_endpoint/requirements.txt
astroid==1.5.3
azure-storage-blob==1.1.0
coverage==4.5.1
Flask==0.12.2
pylint==1.8.1
pytest==3.3.2
pytest-cov==2.5.1
mypy==0.560
requests==2.18.4
setuptools==38.4.0
